requirements.txt
fastapi==0.115.12
sqlalchemy==2.0.41
uvicorn==0.34.3
aiosqlite==0.21.0
alembic==1.16.1
python-dotenv==1.1.0
aiogram==3.20.0.post0
httpx==0.28.1
python-multipart==0.0.20
PyJWT==2.10.1
passlib==1.7.4
